refactored the surgery cost estimation report. report methods were in inwardreportcontroller replaced by surgerycostreportcontroller Signed-off-by: Pubudu <pubudupiyankara.me@gmail.com>
